train-1/RAG/excel/requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
python-dotenv
chromadb
pandas
openpyxl